chore: rename phone_number to phone
